Rationalize usage of integers
  - Except at a few places, int have been replaced by dynamic_graph::size_type,
    also known as Eigen::Matrix::Index.
  - unsigned int have been replaced by std::size_t.

+++ src/tools/integrator.cpp
@@ -188,7 +188,7 @@ Vector &Integrator::integrate(Vector &configuration, sigtime_t time) {
        << velocity.transpose() << ".";
     throw std::runtime_error(os.str().c_str());
   }
-  double delta_t = dt * (time - lastComputationTime_);
+  double delta_t = dt * (double)(time - lastComputationTime_);
   ::pinocchio::integrate(*model_, configuration_, delta_t * velocity,
                          configuration);
   configuration_ = configuration;
